refactor(teams): split manager tests for GH727
